ticketing_system/requirements.txt
Django>=4.0,<5.0
openai>=1.0.0
python-dotenv>=1.0.0
# Remove these if not using local models:
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99
protobuf>=3.20.0
accelerate>=0.20.0
